Beer strength now displays double value

+++ app/src/main/java/com/example/drinksahoy/Beer.kt
@@ -10,7 +10,7 @@ import java.io.Serializable
 data class Beer(
     val imageUrl: String? = null,
     val name: String? = null,
-    val strength: Int? = null,
+    val strength: Double? = null,
     val tagline: String? = null,
     val description: String? = null,
     val foodPair: String? = null
